init: Add a hook to turn a shutdown into a reboot
Intel bootloaders have not yet been made android friendly.  When the
bootloader is told to shutdown it will shutdown.  Android expects the
bootloader to actually shutdown if there is no charger connected and
otherwise to enter charging mode.

Add a hook to init that allows healthd to determine what a shutdown should
do.  When the property is specified, any request to shutdown will instead
trigger a restart into the specified mode.

Change-Id: Ie96fb58ecb9a4e8ffe7e73e590483b5323e40b74