init: Don't crash the system due to an invalid mkdir argument
Prior to this change, if someone write

      mkdir /data/log/sar_cap system system

in one of their init.rc files, this causes init to get a exception
which causes init to exit with no error message to allow you to track
this down which then causes your kernel to panic.

Let's make this a little more user friendly by simply reporting
an error instead of silently crashing the whole damn system.

Change-Id: I2f8c50391e78a9992d33754cf1f1e34e9fcee047
Signed-off-by: Simao Gomes Viana <xdevs23@outlook.com>